Move LDAP stores to ldap folder
