+++ graylog2-web-interface/src/stores/ldap/LdapStore.js
